requirements.txt
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
